improve auth mock
